\name{ONRprocessing-package}
\alias{ONRprocessing-package}
\alias{ONRprocessing}
\docType{package}
\title{
\packageTitle{ONRprocessing}
}
\description{
\packageDescription{ONRprocessing}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ONRprocessing}
\packageIndices{ONRprocessing}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ONRprocessing}

Maintainer: \packageMaintainer{ONRprocessing}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
